+++ src/form-fields/color-picker/presetColors.js
@@ -1,0 +1,86 @@
+// Adopted from Maintenance App Beta
+// https://github.com/dhis2/maintenance-app-beta/blob/master/src/components/ColorAndIconPicker/availableColors.ts
+
+// Flattened array of colors organized by color family for the SketchPicker
+export const AVAILABLE_COLORS = [
+    // Reds
+    '#ffcdd2',
+    '#e57373',
+    '#d32f2f',
+    // Pinks
+    '#f06292',
+    '#f50057',
+    '#c2185b',
+    '#880e4f',
+    // Purples
+    '#e1bee7',
+    '#ba68c8',
+    '#8e24aa',
+    '#aa00ff',
+    // Deep Purples
+    '#7e57c2',
+    '#7c4dff',
+    '#6200ea',
+    '#4527a0',
+    // Blues
+    '#c5cae9',
+    '#7986cb',
+    '#3949ab',
+    '#304ffe',
+    // Light Blues
+    '#e3f2fd',
+    '#64b5f6',
+    '#40c4ff',
+    '#1976d2',
+    '#0288d1',
+    '#00b0ff',
+    // Cyans
+    '#80deea',
+    '#00acc1',
+    '#00838f',
+    '#006064',
+    // Teals
+    '#e0f2f1',
+    '#80cbc4',
+    '#64ffda',
+    '#00695c',
+    // Greens
+    '#c8e6c9',
+    '#66bb6a',
+    '#00e676',
+    '#2e7d32',
+    '#1b5e20',
+    // Light Greens
+    '#aed581',
+    '#76ff03',
+    '#689f38',
+    '#33691e',
+    // Lime/Yellow-Green
+    '#cddc39',
+    '#64dd17',
+    '#9e9d24',
+    '#827717',
+    // Yellows
+    '#fff9c4',
+    '#ffff00',
+    '#fbc02d',
+    '#f57f17',
+    // Oranges/Peach
+    '#ffccbc',
+    '#ffcc80',
+    '#ffab91',
+    // Browns
+    '#bcaaa4',
+    '#8d6e63',
+    '#4e342e',
+    // Grays
+    '#fafafa',
+    '#bdbdbd',
+    '#757575',
+    '#424242',
+    // Blue Grays
+    '#cfd8dc',
+    '#b0bec5',
+    '#607d8b',
+    '#37474f',
+]